Fixing fp64 vs fp32 not equal

+++ tests/test_list/basic_OSERIT_tests/basic_test_019.py
@@ -45,7 +45,7 @@ try:
     if result.variables['lon'][-1][0] <= default_lon:
         error += "particle should go to East"
     
-    if result.variables['lat'][-1][0] != default_lat:
+    if result.variables['lat'][-1][0] != np.float32(default_lat):
         error += "particle should stay at its latitude"
 
     #to the North anticlockwise
@@ -57,7 +57,7 @@ try:
 
     result = run_oserit(id, path_ose, path_tmp, show_ose_res)
 
-    if result.variables['lon'][-1][0] != default_lon:
+    if result.variables['lon'][-1][0] != np.float32(default_lon):
         error += "particle should stay at its longitude"
     
     if result.variables['lat'][-1][0] <= default_lat:
@@ -75,7 +75,7 @@ try:
 
     result = run_oserit(id, path_ose, path_tmp, show_ose_res)
 
-    if result.variables['lon'][-1][0] != default_lon:
+    if result.variables['lon'][-1][0] != np.float32(default_lon):
         error += "particle should stay at its longitude"
     
     if result.variables['lat'][-1][0] >= default_lat:
@@ -90,7 +90,7 @@ try:
 
     result = run_oserit(id, path_ose, path_tmp, show_ose_res)
 
-    if result.variables['lon'][-1][0] != default_lon:
+    if result.variables['lon'][-1][0] != np.float32(default_lon):
         error += "particle should stay at its longitude"
     
     if result.variables['lat'][-1][0] >= default_lat:
